pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openhands-local-agent"
version = "0.1.0"
description = "A local extensible AI agent that uses the OpenAI SDK with Ollama and can operate a browser and terminal."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "playwright>=1.44.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
openhands-agent = "openhands_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
